# server/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: WAPs QR 출석 API
  version: 1.0.0
  description: |
    QR 출석 기능의 API 명세입니다.

    관리자는 출석체크를 생성하고, 진행 중인 출석의 QR을 발급하여 현장에 표시합니다.
    사용자는 로그인한 상태에서 QR을 스캔하여 본인의 출석을 등록하고, 종료된 출석 기록을 조회합니다.
    관리자는 출석·미출석 인원과 명단을 확인하고 출석 상태 및 비고를 수정할 수 있습니다.
    회원 등급에 따른 필드, 필터, 정렬, 출석 제한은 사용하지 않습니다.

    설계 기준:
    - 출석 대상은 생성 시점에 가입된 전체 사용자이며, 대상 명단은 생성 후 고정됩니다.
      모든 대상자의 초기 상태는 ABSENT(미출석)입니다.
    - 출석 날짜는 date 하나로 받으며, 형식은 YYYY-MM-DD입니다.
      한국 시간(Asia/Seoul)을 기준으로 해당 날짜 00:00부터 다음 날 00:00 전까지 출석할 수 있습니다.
    - 한국 시간의 현재 날짜가 date 이전이면 SCHEDULED, date와 같으면 ONGOING,
      date 이후이면 ENDED입니다. 날짜에 따라 자동으로 시작·종료합니다.
    - 관리자 API는 ADMIN 권한, 사용자 API는 로그인과 해당 출석의 대상자 여부를 확인합니다.
    - QR에는 사용자 정보나 로그인 JWT를 넣지 않습니다. 서버가 발급한 출석 전용 토큰만 사용합니다.
servers:
  - url: /
    description: Swagger UI에 접속한 서버
security:
  - JWT: []
tags:
  - name: 관리자 출석
    description: 출석체크 생성, 현황·결과 조회, 수동 수정 및 QR 발급 (ADMIN)
  - name: 사용자 출석
    description: 진행 중인 출석·나의 출석 기록 조회 및 QR 출석 등록 (로그인 필요)
paths:
  /admin/attendances:
    get:
      tags: [관리자 출석]
      operationId: listAdminAttendances
      summary: 출석체크 목록 조회
      description: |
        상태별 출석체크와 출석·미출석 인원을 조회합니다.
        진행 중인 목록은 status=ONGOING, 종료된 목록은 status=ENDED로 요청합니다.
        status를 생략하면 예정된 출석을 포함한 전체 목록을 반환합니다.
        date 내림차순, 같은 날짜이면 attendanceId 내림차순으로 정렬합니다.
        조건에 맞는 전체 목록을 반환하며, 결과가 없으면 content는 빈 배열입니다.
      parameters:
        - name: status
          in: query
          description: 출석체크 진행 상태
          schema:
            $ref: '#/components/schemas/AttendanceStatus'
      responses:
        '200':
          description: 출석체크 목록
          content:
            application/json:
              schema:
                type: object
                required: [content]
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AttendanceSummary'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [관리자 출석]
      operationId: createAttendance
      summary: 출석체크 생성
      description: |
        행사명과 출석 날짜를 지정합니다. 생성 시 전체 사용자를 출석 대상자로 등록합니다.
        title은 앞뒤 공백 제거 후 1~100자여야 하며, date는 한국 시간 기준 오늘 또는 미래 날짜여야 합니다.
        오늘 날짜로 생성하면 즉시 ONGOING 상태가 됩니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttendanceRequest'
      responses:
        '201':
          description: 출석체크 생성 완료
          headers:
            Location:
              description: 생성한 출석체크 상세 조회 경로
              schema:
                type: string
                example: /admin/attendances/1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceSummary'
              example:
                attendanceId: 1
                title: 시작발표
                date: '2026-10-10'
                status: SCHEDULED
                totalCount: 30
                presentCount: 0
                absentCount: 30
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/attendances/{attendanceId}:
    parameters:
      - $ref: '#/components/parameters/AttendanceId'
    get:
      tags: [관리자 출석]
      operationId: getAttendanceDetail
      summary: 출석 현황·결과 및 대상자 명단 조회
      description: |
        진행 중인 출석의 관리 화면과 종료된 출석의 결과 화면에서 사용합니다.
        전체 대상자의 출석·미출석 인원과 대상자별 이름, 출석 상태, 출석 시각, 비고를 반환합니다.
        조건에 맞는 전체 명단을 반환하며, 결과가 없으면 participants.content는 빈 배열입니다.
        인원 집계는 출석 상태 필터와 무관하게 전체 대상자를 기준으로 합니다.
        status 오름차순은 PRESENT 먼저, 내림차순은 ABSENT 먼저입니다.
        정렬 값이 같으면 userId 오름차순으로 정렬합니다.
      parameters:
        - name: status
          in: query
          description: 명단의 출석 상태 필터. 생략하면 전체 대상자를 조회합니다.
          schema:
            $ref: '#/components/schemas/PresenceStatus'
        - name: sort
          in: query
          description: 명단 정렬 기준과 방향
          schema:
            type: string
            enum: ['userName,asc', 'userName,desc', 'status,asc', 'status,desc']
            default: 'userName,asc'
      responses:
        '200':
          description: 출석체크 요약 및 대상자 명단
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AttendanceSummary'
                  - type: object
                    required: [participants]
                    properties:
                      participants:
                        type: object
                        required: [content]
                        properties:
                          content:
                            type: array
                            items:
                              $ref: '#/components/schemas/AttendanceParticipant'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/attendances/{attendanceId}/users/{userId}:
    parameters:
      - $ref: '#/components/parameters/AttendanceId'
      - name: userId
        in: path
        required: true
        description: 출석 대상자의 사용자 ID
        schema:
          type: integer
          format: int64
          minimum: 1
          example: 10
    patch:
      tags: [관리자 출석]
      operationId: updateAttendanceParticipant
      summary: 대상자 출석 상태·비고 수정
      description: |
        진행 중이거나 종료된 출석의 상태 또는 비고를 수정합니다. 예정된 출석은 수정할 수 없습니다.
        status와 note 중 하나 이상을 전송해야 하며, 생략한 필드는 기존 값을 유지합니다.
        ABSENT에서 PRESENT로 변경하면 checkedInAt은 서버의 현재 시각, ABSENT로 변경하면 null입니다.
        같은 status를 다시 전송하면 기존 checkedInAt을 유지합니다. 빈 문자열 note는 비고를 지웁니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                status:
                  $ref: '#/components/schemas/PresenceStatus'
                note:
                  type: string
                  maxLength: 500
                  example: QR 인식 오류로 관리자 확인 후 출석 처리
      responses:
        '200':
          description: 수정된 출석 대상자 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceParticipant'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 아직 시작되지 않은 출석체크 (COMMON_CONFLICT)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2026-10-09T10:00:00Z'
                status: 409
                code: COMMON_CONFLICT
                message: 아직 시작되지 않은 출석체크입니다.
                path: /admin/attendances/1/users/10
  /admin/attendances/{attendanceId}/qr:
    parameters:
      - $ref: '#/components/parameters/AttendanceId'
    post:
      tags: [관리자 출석]
      operationId: issueAttendanceQr
      summary: 출석 QR 발급·재발급
      description: |
        ONGOING인 출석체크의 QR 토큰을 발급합니다.
        프런트엔드는 응답의 attendanceId와 qrToken을 JSON 문자열로 직렬화하여 QR 이미지로 표시합니다.
        예: {"attendanceId":1,"qrToken":"example-attendance-qr-token"}
        QR 이미지 생성 및 카메라 스캔은 프런트엔드에서 처리합니다.

        토큰은 추측 불가능한 출석 전용 난수이며, 유효 시간은 발급 시각부터 30초입니다.
        출석 날짜가 지나면 토큰의 유효 시간과 무관하게 출석을 등록할 수 없습니다.
        재발급 시 해당 출석체크의 이전 토큰은 즉시 무효화됩니다.
        동일 토큰을 여러 대상자가 사용할 수 있지만, 사용자별 출석은 한 번만 기록합니다.
        응답에는 Cache-Control: no-store를 적용합니다.
      responses:
        '200':
          description: QR로 표시할 출석체크 ID, 토큰 및 만료 시각
          headers:
            Cache-Control:
              schema:
                type: string
                enum: [no-store]
          content:
            application/json:
              schema:
                type: object
                required: [attendanceId, qrToken, expiresAt]
                properties:
                  attendanceId:
                    type: integer
                    format: int64
                    minimum: 1
                    example: 1
                  qrToken:
                    type: string
                    description: 로그인 JWT와 별개인 출석 전용 토큰
                    example: example-attendance-qr-token
                  expiresAt:
                    type: string
                    format: date-time
                    description: 발급 시각으로부터 30초 후의 시각. 서버 시각이 이 시각 이상이면 토큰은 만료됩니다.
                    example: '2026-10-10T19:00:30+09:00'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/AttendanceNotOngoing'
  /attendances:
    get:
      tags: [사용자 출석]
      operationId: listMyAttendances
      summary: 진행 중인 출석·나의 출석 기록 조회
      description: |
        로그인 사용자가 대상자인 출석체크와 본인의 출석 상태를 반환합니다.
        진행 중인 출석은 status=ONGOING, 종료된 나의 출석 기록은 status=ENDED로 요청합니다.
        status를 생략하면 ONGOING으로 조회합니다.
        사용자 ID는 JWT에서 확인하며 요청 파라미터로 받지 않습니다.
        date 내림차순, 같은 날짜이면 attendanceId 내림차순으로 정렬합니다.
        조건에 맞는 전체 목록을 반환하며, 결과가 없으면 빈 배열입니다.
      parameters:
        - name: status
          in: query
          description: 진행 중인 출석 또는 종료된 나의 출석 기록 선택
          schema:
            type: string
            enum: [ONGOING, ENDED]
            default: ONGOING
      responses:
        '200':
          description: 출석체크 정보와 본인의 출석 상태를 포함한 카드 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/Attendance'
                    - type: object
                      required: [myStatus, checkedInAt]
                      properties:
                        myStatus:
                          $ref: '#/components/schemas/PresenceStatus'
                        checkedInAt:
                          $ref: '#/components/schemas/CheckedInAt'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /attendances/{attendanceId}/check-in:
    parameters:
      - $ref: '#/components/parameters/AttendanceId'
    post:
      tags: [사용자 출석]
      operationId: checkInAttendance
      summary: QR 스캔으로 본인 출석 등록
      description: |
        스캔한 QR의 attendanceId를 경로에, qrToken을 요청 본문에 넣습니다.
        서버는 JWT의 사용자 ID로 본인의 출석만 처리하며 타인의 userId는 받지 않습니다.
        출석체크가 ONGOING이고 본인이 대상자이며, 해당 출석체크의 최신 QR 토큰이 유효해야 합니다.

        최초 성공 시 PRESENT로 변경하고 서버의 현재 시각을 checkedInAt에 기록합니다.
        유효한 QR로 이미 출석한 사용자가 다시 요청하면 200과 기존 출석 시각을 반환합니다.
        동시 요청에도 (attendanceId, userId)별 출석은 한 번만 기록하며 집계가 중복 증가하지 않습니다.
        QR 출석은 관리자가 입력한 비고를 변경하지 않습니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [qrToken]
              properties:
                qrToken:
                  type: string
                  minLength: 1
                  maxLength: 512
                  example: example-attendance-qr-token
      responses:
        '200':
          description: 본인 출석 완료 또는 이미 출석한 사용자의 기존 결과
          content:
            application/json:
              schema:
                type: object
                required: [attendanceId, userId, status, checkedInAt]
                properties:
                  attendanceId:
                    type: integer
                    format: int64
                    minimum: 1
                    example: 1
                  userId:
                    type: integer
                    format: int64
                    minimum: 1
                    example: 10
                  status:
                    type: string
                    enum: [PRESENT]
                    example: PRESENT
                  checkedInAt:
                    type: string
                    format: date-time
                    example: '2026-10-10T19:00:15+09:00'
        '400':
          description: |
            잘못된 입력 또는 유효하지 않은 QR (COMMON_INVALID_INPUT).
            토큰 위조, 만료, 재발급으로 무효화된 토큰, 다른 출석체크의 토큰을 포함합니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2026-10-10T10:00:31Z'
                status: 400
                code: COMMON_INVALID_INPUT
                message: QR이 유효하지 않거나 만료되었습니다. 최신 QR을 다시 스캔해주세요.
                path: /attendances/1/check-in
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/AttendanceNotOngoing'
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 기존 로그인 API에서 발급받은 액세스 토큰. QR 토큰과는 다릅니다.
  parameters:
    AttendanceId:
      name: attendanceId
      in: path
      required: true
      description: 출석체크 ID
      schema:
        type: integer
        format: int64
        minimum: 1
        example: 1
  schemas:
    AttendanceStatus:
      type: string
      enum: [SCHEDULED, ONGOING, ENDED]
      description: 예정 / 진행 중 / 종료. 출석 날짜와 한국 시간 기준 현재 날짜로 결정합니다.
      example: ONGOING
    PresenceStatus:
      type: string
      enum: [PRESENT, ABSENT]
      description: 출석 / 미출석. 지각은 별도 상태로 구분하지 않고 PRESENT(출석)로 간주합니다.
      example: ABSENT
    CheckedInAt:
      type: string
      format: date-time
      nullable: true
      description: 출석 처리 시각. 미출석 상태이면 null입니다.
      example: null
    CreateAttendanceRequest:
      type: object
      additionalProperties: false
      required: [title, date]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          description: 앞뒤 공백을 제거한 행사명. 공백만 입력할 수 없습니다.
          example: 시작발표
        date:
          type: string
          format: date
          description: 한국 시간 기준 오늘 또는 미래의 출석 날짜 (YYYY-MM-DD)
          example: '2026-10-10'
    Attendance:
      type: object
      required: [attendanceId, title, date, status]
      properties:
        attendanceId:
          type: integer
          format: int64
          minimum: 1
          example: 1
        title:
          type: string
          example: 시작발표
        date:
          type: string
          format: date
          description: 한국 시간 기준 출석 날짜 (YYYY-MM-DD)
          example: '2026-10-10'
        status:
          $ref: '#/components/schemas/AttendanceStatus'
    AttendanceSummary:
      allOf:
        - $ref: '#/components/schemas/Attendance'
        - type: object
          required: [totalCount, presentCount, absentCount]
          properties:
            totalCount:
              type: integer
              minimum: 0
              description: 전체 대상자 수. presentCount + absentCount와 같습니다.
              example: 30
            presentCount:
              type: integer
              minimum: 0
              example: 20
            absentCount:
              type: integer
              minimum: 0
              example: 10
    AttendanceParticipant:
      type: object
      required: [userId, userName, status, checkedInAt, note]
      properties:
        userId:
          type: integer
          format: int64
          minimum: 1
          example: 10
        userName:
          type: string
          example: 잔망루피
        status:
          $ref: '#/components/schemas/PresenceStatus'
        checkedInAt:
          $ref: '#/components/schemas/CheckedInAt'
        note:
          type: string
          maxLength: 500
          description: 관리자 비고. 비고가 없으면 빈 문자열입니다.
          example: ''
    ErrorResponse:
      type: object
      description: 기존 서버의 공통 ErrorResponse 형식을 따릅니다. 비어 있는 선택 필드는 생략합니다.
      required: [timestamp, status, code, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2026-10-10T10:00:00Z'
        status:
          type: integer
          example: 400
        code:
          type: string
          example: COMMON_INVALID_INPUT
        message:
          type: string
          example: 잘못된 요청입니다.
        path:
          type: string
          example: /admin/attendances
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
                example: date
              message:
                type: string
                example: 출석 날짜는 한국 시간 기준 오늘 또는 미래 날짜여야 합니다.
              rejectedValue:
                nullable: true
                description: 검증에 실패한 입력값. 문자열, 숫자, 객체 또는 null 등 원래 타입을 유지합니다.
        requestId:
          type: string
          description: 서버에서 제공하는 경우에만 포함되는 요청 추적 ID
  responses:
    InvalidInput:
      description: 입력값, 날짜 또는 정렬 조건이 잘못됨 (COMMON_INVALID_INPUT)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 로그인 토큰 누락·오류·만료 (AUTH_UNAUTHORIZED / AUTH_INVALID_TOKEN / AUTH_TOKEN_EXPIRED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-10-10T10:00:00Z'
            status: 401
            code: AUTH_UNAUTHORIZED
            message: 인증이 필요합니다.
            path: /admin/attendances
    Forbidden:
      description: 관리자 권한이 없거나 해당 출석체크의 대상자가 아님 (AUTH_FORBIDDEN)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-10-10T10:00:00Z'
            status: 403
            code: AUTH_FORBIDDEN
            message: 접근 권한이 없습니다.
            path: /admin/attendances
    NotFound:
      description: 출석체크 또는 출석 대상자를 찾을 수 없음 (COMMON_RESOURCE_NOT_FOUND)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-10-10T10:00:00Z'
            status: 404
            code: COMMON_RESOURCE_NOT_FOUND
            message: 요청한 출석체크 또는 출석 대상자를 찾을 수 없습니다.
            path: /admin/attendances/1
    AttendanceNotOngoing:
      description: 예정되었거나 종료되어 출석을 진행할 수 없음 (COMMON_CONFLICT)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-10-11T00:00:00Z'
            status: 409
            code: COMMON_CONFLICT
            message: 진행 중인 출석체크가 아닙니다.
            path: /attendances/1/check-in